requirements.txt
# ==========================================
# MoodTracker - Comprehensive Multi-Modal Emotion Analysis Application
# All dependencies consolidated from entire project
# ==========================================
# To install all dependencies:
#   pip install -r requirements.txt
# ==========================================

# ===== CORE WEB FRAMEWORK =====
streamlit>=1.28.1
streamlit-webrtc>=0.47.0

# ===== COMPUTER VISION & FACE DETECTION =====
opencv-contrib-python-headless>=4.8.0.74
av>=10.0.0

# ===== DEEP LEARNING & NEURAL NETWORKS =====
keras>=2.10.0,<3.0.0
tensorflow-cpu>=2.14.0,<2.22.0
h5py>=3.7.0
protobuf>=3.13.0,<5.0.0
Keras-Applications>=1.0.8
Keras-Preprocessing>=1.1.2
tensorboard>=2.1.1
tensorflow-estimator>=2.1.0
opt-einsum>=3.3.0
google-pasta>=0.2.0

# ===== DATA PROCESSING & NUMERICAL COMPUTING =====
numpy>=1.26.0
pandas>=2.0.3
scipy>=1.4.1
scikit-learn>=1.3.0
joblib>=1.3.1

# ===== NATURAL LANGUAGE PROCESSING =====
textblob>=0.17.1
nltk>=3.7
neattext>=0.1.0
regex>=2022.9.13

# ===== SPEECH & AUDIO PROCESSING =====
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
pydub>=0.25.1
sounddevice>=0.4.5
ffmpeg-python>=0.2.1

# ===== DATA VISUALIZATION =====
matplotlib>=3.7.1
altair>=5.0.1
plotly>=5.0.0
seaborn>=0.12.0

# ===== IMAGE & FILE HANDLING =====
Pillow>=9.2.0

# ===== DATABASE & ORM =====
SQLAlchemy>=1.4.41

# ===== SECURITY & CRYPTOGRAPHY =====
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycryptodome>=3.15.0
rsa>=4.6

# ===== UTILITIES & HELPERS =====
python-dateutil>=2.8.2
pytz>=2020.1
PyYAML>=6.0
pyzmq>=24.0.1
requests>=2.24.0
requests-oauthlib>=1.3.1
pyparsing>=3.0.9

# ===== OPTIONAL: MEDICAL IMAGING & COMPUTER VISION =====
pydicom>=2.3.0
pycocotools>=2.0.5

# ===== WEB FRAMEWORKS =====
Django>=3.1.2
Flask>=1.1.2

# ===== ADDITIONAL UTILITIES & SUPPORT =====
beautifulsoup4>=4.11.1
tqdm>=4.64.1
click>=7.1.2
grpcio>=1.49.1

# ===== SECURITY & AUTHENTICATION =====
google-auth>=1.23.0
google-auth-oauthlib>=0.4.6
certifi>=2020.6.20

# ===== RUNNING THE APPLICATION =====
# Main Hub (Landing Page):
#   streamlit run landing_page.py
#
# The app includes 3 emotion analysis features:
#   1. NLP Text Emotion - Analyze text sentiment
#   2. Real-Time Detection - Webcam-based emotion detection  
#   3. Voice Mood Analyzer - Record/upload audio for mood analysis
# ==========================================

# Made by: Aditya Sharma
# © 2026 MoodTracker. All rights reserved.
